Agregar Linea?

+++ OnlineCarRentalSystem/src/EntidadesOCR/classes/DTOresumen.java
@@ -16,8 +16,8 @@ public class DTOresumen {
     private String mensaje;
     private Collection<Linea> lineas;
     private double total_renta;
-    private int saldo_ingresados;
-    private int cantidad_vueltos;
+    private double saldo_ingresados;
+    private double cantidad_vueltos;
 
     public String getMensaje() {
         return mensaje;
@@ -43,19 +43,19 @@ public class DTOresumen {
         this.total_renta = total_renta;
     }
 
-    public int getSaldo_ingresados() {
+    public double getSaldo_ingresados() {
         return saldo_ingresados;
     }
 
-    public void setSaldo_ingresados(int saldo_ingresados) {
+    public void setSaldo_ingresados(double saldo_ingresados) {
         this.saldo_ingresados = saldo_ingresados;
     }
 
-    public int getCantidad_vueltos() {
+    public double getCantidad_vueltos() {
         return cantidad_vueltos;
     }
 
-    public void setCantidad_vueltos(int cantidad_vueltos) {
+    public void setCantidad_vueltos(double cantidad_vueltos) {
         this.cantidad_vueltos = cantidad_vueltos;
     }
 
@@ -69,7 +69,7 @@ public class DTOresumen {
         return ret;
     }
 
-    public DTOresumen(Collection<Linea> lineas, double total_renta, int saldo_ingresados, int cantidad_vueltos) {
+    public DTOresumen(Collection<Linea> lineas, double total_renta, double saldo_ingresados, double cantidad_vueltos) {
         this.lineas = lineas;
         this.total_renta = total_renta;
         this.saldo_ingresados = saldo_ingresados;
